# projects/restful-admin-crud/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API 제공 HUB 입니다.",
        "title": "API-HUB",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api-hub.ppeum.com",
    "basePath": "/",
    "paths": {
        "/v1/admin/permission-approvals": {
            "patch": {
                "description": "관리자가 권한 요청을 승인 또는 거절합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "권한 승인 요청"
                ],
                "summary": "권한 요청 승인 또는 거절",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003ctoken\u003e",
                        "description": "Bearer JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "권한 승인 또는 거절 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PermissionApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 형식",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse400"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse500"
                        }
                    }
                }
            }
        },
        "/v1/admin/permission-approvals-requests": {
            "get": {
                "description": "페이징을 지원하여 권한 요청 목록을 검색합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "권한 승인 요청"
                ],
                "summary": "권한 승인 요청 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "페이지 번호 (기본값: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지당 항목 수 (기본값: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PermissionApprovalsRequestsPagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 형식",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse400"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.PermissionApprovalRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "default": "approved",
                    "enum": [
                        "approved",
                        "rejected"
                    ],
                    "example": "approved"
                }
            }
        },
        "models.PermissionApprovalsRequests": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "requested_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.PermissionApprovalsRequestsPagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PermissionApprovalsRequests"
                    }
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "`omitempty`: 값이 없을 경우 JSON에서 제외"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}